third_party/blink/web_tests/virtual/fenced-frame-mparch/wpt_internal/fenced_frame/resize-lock.https-expected.txt

This is a testharness.js-based test.
FAIL Test Resize Lock assert_equals: fenced frame dimensions should not be updated by parent page expected "300x150" but got "444x444"
Harness: the test ran to completion.
